Sample application shoing log4cocoa's use. Changed raising NSException to just log a base error instead. Changed tests to not use unsigned values, but to typecast instead. Added a precompiled version of OCMock. Added xcconfig files for many differnt ways to build.
git-svn-id: https://log4cocoa.svn.sourceforge.net/svnroot/log4cocoa/trunk/source@111 3ad47a68-a840-0410-8984-94a09cb05a02

+++ Tests/TestL4Properties.m
@@ -46,7 +46,7 @@
 	
 	L4Properties *properties = [L4Properties propertiesWithFileName:filename];
 	STAssertNotNil(properties, @"L4Properties could not be instantiated.");
-	STAssertEquals([properties count], 10u, @"Expected 10 keys, but found %i", [properties count]);
+	STAssertEquals([properties count], (NSUInteger)10, @"Expected 10 keys, but found %i", [properties count]);
 	
 	STAssertEqualObjects([properties stringForKey:@"log4cocoa.appender.A1"], @"L4ConsoleAppender", 
 						 @"Expected 'L4ConsoleAppender', but found %@", 
@@ -87,7 +87,7 @@
 
 	L4Properties *properties = [L4Properties propertiesWithProperties:fullDictionary];
 	STAssertNotNil(properties, @"L4Properties could not be instantiated.");
-	STAssertEquals([properties count], 10u, @"Expected 10 keys, but found %i", [properties count]);
+	STAssertEquals([properties count], (NSUInteger)10, @"Expected 10 keys, but found %i", [properties count]);
 	
 	STAssertEqualObjects([properties stringForKey:@"log4cocoa.appender.A1"], @"L4ConsoleAppender", 
 						 @"Expected 'L4ConsoleAppender', but found %@",
